codalab/requirements/common.txt
Django==1.11.29
alembic==0.6.7
anyjson==0.3.3
argparse
# azure==0.7.1
closure_linter
django-analytical==2.0.0
django-appconf==1.0
django-compressor==2.2
django-configurations==2.0
django-debug-toolbar==2.2
sqlparse==0.2.0
django-debug-panel==0.8.1
django-extensions==1.7.6
django-extra-views==0.6.3
django-filter==1.1.0
django-guardian==1.4.9
django-haystack==2.8.1
django-js-reverse>=0.2.0
django-mptt==0.9.1
django-nose==1.4.6
django-oauth-toolkit==0.11.0
django-storages==1.9.1
django-tinymce==2.7.1
djangorestframework==3.8.2
mock==1.0.1
oauthlib==1.1.2
psutil==5.7.2
python-openid==2.2.5
python-ptrace==0.6.5
pytz==2015.7
PyYAML==5.1
requests-oauthlib==0.3.2
requests==2.24.0
six==1.10.0
SQLAlchemy==0.9.7
watchdog==0.8.1
#wsgiref==0.1.2
django-simple-captcha==0.5.12
django-bootstrap-form==3.4
newrelic==2.54.0.41
pylibmc==1.5.2

celery==4.4.6
boto3==1.14.22
# We need a modified s3direct 0.4.1 which can use AWS_S3_HOST instead of forcing aws
git+https://github.com/codalab/django-s3direct.git@feature/allow-different-endpoint

#pyrabbit==1.1.0
git+https://github.com/bkjones/pyrabbit.git@e8a9f74ed5c6bba958994fb9a72c396e6a99ea0f

gunicorn==20.0.4
#gevent==1.2.2

#Extras added
git+https://github.com/Tthomas63/django-switchuser.git@convert-python3
Sphinx==1.4.1

# Patched a vulnerability in allauth. The login view redirects to the next parameter in the url after successful form validation. Before, there was
# minimal validation, so it could be manually changed in the browser to any host. Now, it is checked against django.utils.is_safe_url before the 
# redirect occurs.
# 7/16/2020: I would think the above vulnerability would be patched in later versions that we are using now. Leaving in-case we need to reference
# git+https://github.com/gibsonbailey/django-allauth.git@codalab
django-allauth==0.35.0

# Added for python sdk
git+https://github.com/FlavioAlexander/azure-sdk-for-python.git@legacy

# Required when PostgreSQL is used
psycopg2==2.8.3

# Lock unwanted people out on staging servers
django-pin-passcode==0.1.9

# Adding test libraries here for simplicity sake
pytest==5.4.3
pytest-django==3.9.0
pytest-testmon==1.0.2
# I haven't found any uses of pytest-watch, but it seems like something that should be leveraged when I looked into it.
pytest-watch==4.2.0
pytest-env==0.6.2

lxml==4.1.1

certifi==2020.6.20
